lsst-build: initial implementation
With --version-git-repo set to point to a specially formatted git repository,
+SHA1 versions will be prettified by replacing the hashes with +N numbers,
the mapping to which is kept in the repository.

The repository also keeps the mapping of +N -> version of dependencies, as
well as the build manifests for each build.

In lieu of real documentation, there's doc/rebuild.sh for now.