add --repos argument
The option --repos argument takes a path to a YAML formatted file of a
hash.  The hash keys match to EUPS products names and the values are the
URL to said product's git repo. Eg.

    ---
    afw: https://github.com/lsst/afw.git
    qserv: https://github.com/lsst/qserv.git

If --repos is specified in combination with either the --repository-pattern
option or REPOSITORY_PATTERN environment variable, the repos provided URL
is tested as a working remote first before falling back to the
--repository-pattern behavior.

Out of necessity in integrating this feature, the default string for
--repository--pattern has been removed.  This was necessary so that --repos
could be provided without any fall back patterns to enforce the origin
remote to be updated in any existing clones.