Support post-clone enabling of Git LFS.
Previously, we only enabled Git LFS for a repository when cloning it. If
it was subsequently enabled in repos.yaml, that change would not propagate to
our clone.

This changes things so that the clone and the check for LFS are separate. When
LFS is enabled in repos.yaml, already cloned repositories will have it
switched on at their next update.